pyproject.toml
[project]
name = "agent-engine-custom-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-aiplatform[adk,agent-engines]>=1.94.0",
    "google-genai>=1.16.1",
    "ruff>=0.11.11",
]
